==> 1_raw_corpus.ipynb <==
# Raw Rigveda Corpus

## Getting Rigveda samhitapatha (VNH version)

# https://github.com/cceh/c-salt_vedaweb_sources/tree/master/rigveda/versions
# description of the sources here:
# https://github.com/cceh/c-salt_vedaweb_tei/blob/master/vedaweb_corpus.tei
# https://vedaweb.uni-koeln.de/rigveda/help

!wget -O downloads/rv_samhitapatha_vnh.json https://raw.githubusercontent.com/cceh/c-salt_vedaweb_sources/master/rigveda/versions/vnh.json

# make text version from the jsons, with line numbers at the beginning
!python src/transform_json_corpus.py downloads/rv_samhitapatha_vnh.json

Make sure that the missing chars here are okay to ignore, or that they are just written differently in the text.

# no of verse lines
!cat data/rv_samhitapatha_vnh.txt | awk NF | wc -l

## Getting Rigveda padapatha (Lubotsky version)

!wget -O downloads/rv_padapatha_lubotsky.json https://raw.githubusercontent.com/cceh/c-salt_vedaweb_sources/master/rigveda/versions/lubotsky.json

# make text version from the jsons, with line numbers at the beginning
!python src/transform_json_corpus.py downloads/rv_padapatha_lubotsky.json

Make sure that the missing chars here are okay to ignore, or that they are just written differently in the text.

# no of verse lines
!cat data/rv_padapatha_lubotsky.txt | awk NF | wc -l

The slight difference in no. of verse lines here with that of the VNH samhitapatha is due to how the differences in how the line division is done in the padapatha vs samhitapatha, for certain stanzas that don't follow a standard meter. 

## Getting Rigveda padapatha (Eichler version)

# http://www.detlef108.de/Rigveda.htm 
# http://www.detlef108.de/Notes-to-the-Rigveda-Page.htm 
!wget -O downloads/rv_padapatha_eichler.html http://www.detlef108.de/RV-Padapatha-TA3-paada-NA-UTF8.html 

# sudo apt install html2text
#!html2text -utf8 -width 3000 -o rv_padapatha.txt rv_padaptaha.html

from bs4 import BeautifulSoup

with open("downloads/rv_padapatha_eichler.html", "r") as input_file:
    soup = BeautifulSoup(input_file)
    
    hymns = []
    
    for para in soup.find_all("p"):
        # ignore the ending notes
        if para.contents[0].name == "span":
            continue
        
        #hymns.append(para.text.rstrip()) # no extra lines between hymns
        hymns.append(para.text)
    
    with open("data/rv_padapatha_eichler.txt", 'w') as f:
        f.write("".join(hymns))

# TODO break the padapatha verse into sub-lines

## Getting Strata Info

RV strata assignment for each pāda (verse line) of each stanza:

!wget -O downloads/rv_strata_vedaweb.json https://raw.githubusercontent.com/VedaWebProject/vedaweb-data/main/rigveda/info/strata.json

import json

rv_strata_vedaweb = []

with open(f"downloads/rv_strata_vedaweb.json") as f:
    strata_raw = json.load(f)
    
    for stanza_no, verse_lines  in strata_raw.items():
        for line_data in verse_lines:
            rv_strata_vedaweb.append({
                "line_no": f"{stanza_no}.{line_data[0]}",
                "labels": line_data[1], # extra info
                "strata": line_data[2].upper(), # normalize smallcase strata values
                # if false, based on both meter and linguistic evidence
                "strata_based_on_meter_only": line_data[2].islower(),
            })

rv_strata_vedaweb[0] # sample

import pandas

df_rv_strata_vedaweb = pandas.DataFrame.from_dict(rv_strata_vedaweb)
df_rv_strata_vedaweb.to_csv("data/rv_strata_vedaweb.csv", index=None)
#df_rv_strata_vedaweb.head()

df_rv_strata_vedaweb \
    .groupby(["strata_based_on_meter_only", "strata"]) \
    .size().to_frame("count").reset_index()
