==> 2_roots.ipynb <==
# Roots and Stems for Ninth Class (and Fifth Class)

## Getting Verbal Roots 

We rely on the [list of stems](https://www.sanskrit-lexicon.uni-koeln.de/scans/csl-whitroot/disp/index.php?page=214) given in the index of Whitney's Sanskrit Roots and Verb-forms.

!mkdir -p downloads
!mkdir -p data

!wget -O downloads/whitney_roots.pdf http://gretil.sub.uni-goettingen.de/gretil_elib/Whi885__Whitney_Roots-ACCENTED.pdf

# install pdftk if not already there. eg: for ubuntu: sudo apt install pdftk
!pdftk downloads/whitney_roots.pdf cat 229 output data/whitney_roots_ninth_class.pdf

# for our control data
!pdftk downloads/whitney_roots.pdf cat 228 output data/whitney_roots_fifth_class.pdf

# produces data/whitney_roots_ninth_class.txt
!pdftotext data/whitney_roots_ninth_class.pdf

# produces data/whitney_roots_fifth_class.txt
!pdftotext data/whitney_roots_fifth_class.pdf

The text version is cleaned up manually, fixing formatting and diacritics.

One extra thing we also do is rewrite a form like _mī̆nā_ as _minā/mīnā_, i.e. re-write the variation in the root vowel as two different stem forms explicitly. This helps us visualize and process the variants easily later (note: whitney has only 3 stems here marked like this -- namely _mī̆nā_, _vlī̆nā_ and _dhū̆nī_ -- so we can get away with doing this manually here easily. If there were a lot of these, we could have automated it)

Final results are in [data/whitney_roots_ninth_class_cleaned.txt](data/whitney_roots_ninth_class_cleaned.txt) and [data/whitney_roots_fifth_class_cleaned.txt](data/whitney_roots_fifth_class_cleaned.txt).

## Parsing Verbal Roots Info

# in the same folder as this notebook
import src.lib.roots as roots

# useful during testing to pick up changes in the file
import importlib
importlib.reload(roots)

whitney_roots = roots.parse_whitney_roots([
    "data/whitney_roots_ninth_class_cleaned.txt",
    "data/whitney_roots_fifth_class_cleaned.txt",
])

## Saving the Results

import pandas

df_whitney_roots = pandas.DataFrame.from_dict(whitney_roots)
df_whitney_roots.to_csv("data/roots.csv", index=None)

## Summary of Results

import pandas
res_df_roots = pandas.read_csv("data/roots.csv", keep_default_na=False)
res_df_roots

res_df_roots.groupby(["present_class", "language_period"]).size().to_frame("count").reset_index()

For Rigveda, we don't expect to find the roots marked as "Later" in the language period.
