==> 3_roots_with_attestations.ipynb <==
# Roots and Stems With Stanza Attestations

## Enriching list of Verbal Roots with Rigveda Attestations

import pandas

# see 2_roots.ipynb for details on this file
df_roots = pandas.read_csv("data/roots.csv", keep_default_na=False)
#df_roots.head()

roots = df_roots.to_dict("records")
roots[0] # sample

# in the same folder as this notebook
import src.lib.roots_attestations as roots_attestations

# useful during testing to pick up changes in the file
import importlib
importlib.reload(roots_attestations)

# uses the vedaweb api underneath for the searches
roots_with_attestations, roots_with_attested_words = roots_attestations.get_attestations(roots)

Now let's look at some of the sample results.

roots_with_attestations[2] # sample

roots_with_attested_words["ninth"]["ubh"] # sample

## Saving the Results

import json
with open(f"data/roots_with_attested_words.json", 'w') as f:
    json.dump(roots_with_attested_words, f, indent=2, ensure_ascii=False)

import pandas
df_roots_with_attestations = pandas.DataFrame.from_dict(roots_with_attestations)
df_roots_with_attestations.to_csv("data/roots_with_attestations.csv", index=None)
df_roots_with_attestations.head()

## Validation

### Checking for attested roots

# roots attested in RV
df_roots_with_attestations.query(
    'strong_attestations_total > 0 or weak_attestations_total > 0'
).sort_values(["present_class", "language_period", "root"])

TODO explain cases where root_guess and root differ

eg: iS 'send' being marked with 1 automatically from vedaweb (same as what lubotsky gives)
whitney actually has this as 2 in main root list but since stem does not have variants with it, it's not marked later

### Checking for missing roots

# to print without index on the left
#from IPython.display import HTML
#HTML(df_roots_with_attestations.to_html(index=False))

# roots not attested in RV
df_roots_with_attestations.query(
    'strong_attestations_total == 0 and weak_attestations_total == 0'
).sort_values(["present_class", "language_period", "root"])

~TODO for the ninth class, test these with different length of final root vowel? just to see if we catch anything~

Done and it didn't. 

## Summary of Results

import pandas
res_df_roots_with_attestations = pandas.read_csv("data/roots_with_attestations.csv", keep_default_na=False)
res_df_roots_with_attestations.head()

# query for attested roots
res_df_roots_with_attestations \
    .query('strong_attestations != "" or weak_attestations != ""') \
    .groupby(["present_class", "language_period"]) \
    .size().to_frame("count").reset_index()
