==> 4_verse_lines.ipynb <==
# Verse Lines with Roots and Stems

## Organizing Root Attestations Data by Verse Lines (pādas)

import pandas

# see 3_roots_with_attestations.ipynb for details on this file
df_roots = pandas.read_csv("data/roots_with_attestations.csv", keep_default_na=False)
#df_roots.head()

#df_roots_attested = df_roots[~df_roots["strong_attestations"].isna()]
#df_roots_attested = df_roots[df_roots["strong_attestations"].str.len() > 0]
df_roots_attested = df_roots.query('strong_attestations != "" or weak_attestations != ""')

df_roots_attested.head()

rv_lines = []

roots_attested = df_roots_attested.to_dict("records")

for root in roots_attested:
    rv_weak_stanza_nos = root.pop("weak_attestations").split()
    rv_strong_stanza_nos = root.pop("strong_attestations").split()   
    # FIXME just eliminate these? since these are only stanza
    root.pop("weak_attestations_total")
    root.pop("strong_attestations_total")
    
    weak_stem = root.pop("weak_stem")
    strong_stem = root.pop("strong_stem")
    
    # we set line_no to be stanza_no, to start off with
    # (will be updated to actual line_no later)
    for stanza_no in rv_weak_stanza_nos:
        rv_lines.append({"line_no": stanza_no, "stem": weak_stem, "stem_type": "weak"} | root)  
    for stanza_no in rv_strong_stanza_nos: 
        rv_lines.append({"line_no": stanza_no, "stem": strong_stem, "stem_type": "strong"} | root)

rv_lines[0]

## Parsing Line numbers into Constituents

# "1.1.1b" > "01" "001" "02" "b"
# 01.063.02 > "01" "063" "02" ""
def parse_rv_line_no(string):
    line_no_parts = string.split(".")
    
    book = line_no_parts[0].zfill(2)
    hymn = line_no_parts[1].zfill(3)
    
    last_char = line_no_parts[2][-1]
    if last_char.isalpha(): # if pada is already present
        stanza = line_no_parts[2][:-1] # drop the last char
        pada = last_char
    else:
        stanza = line_no_parts[2]
        pada = ""

    stanza = stanza.zfill(2)

    return {
        "book"    : book,
        "hymn"    : f"{book}.{hymn}",
        "stanza"  : f"{book}.{hymn}.{stanza}",
        "pada"    : f"{book}.{hymn}.{stanza}.{pada}" if pada else ""
        #"pada_id" : pada or ''
    }    

rv_lines = [line | (parse_rv_line_no(line["line_no"])) for line in rv_lines]

rv_lines[0]

## Enrching Verse Lines with Stanza Metadata

### Downloading annotation data

!mkdir -p downloads/vedaweb

import os
import requests
import time

VEDAWEB_API_URL = "https://vedaweb.uni-koeln.de/rigveda/api"

rv_stanza_nos = sorted(list(set([line["stanza"] for line in rv_lines])))

# no of stanzas for which to dowload the stanza data
len(rv_stanza_nos)

for stanza_no in rv_stanza_nos:
    print(f"Getting data for stanza: {stanza_no}")
    
    stanza_file = f"downloads/vedaweb/{stanza_no}.json"
    if os.path.exists(stanza_file):
        # we already downloaded this stanza so continue to the next one
        continue
    
    # eg: https://vedaweb.uni-koeln.de/rigveda/api/document/id/0100102
    vedaweb_doc_id = stanza_no.replace('.', '')
    vedaweb_doc_url = f"{VEDAWEB_API_URL}/document/id/{vedaweb_doc_id}"
    
    response = requests.get(vedaweb_doc_url)
    # raises an exception on non-200 responses, since we want to know and act on it
    response.raise_for_status()
    
    with open(stanza_file, 'w') as f:
        f.write(response.text)
    
    # so that we don't hammer the api
    time.sleep(0.5)

print("Done!")

### Enriching the lines with text and metrical info

import json

with open(f"data/roots_with_attested_words.json") as f:
    roots_with_attested_words = json.load(f)

# in the same folder as this notebook
import src.lib.verse_lines as verse_lines

# useful during testing to pick up changes in the file
import importlib
importlib.reload(verse_lines)

rv_lines_annotated = []
for line in rv_lines:
    # we can get multiple pada lines from a singe stanza attestation so need to use extend
    rv_lines_annotated.extend(verse_lines.annotate_line(line, roots_with_attested_words))

print(f"\nTotal number of lines: {len(rv_lines_annotated)}")

rv_lines_annotated[0] # sample

## Saving the Results

df_rv_lines = pandas.DataFrame.from_dict(rv_lines_annotated)
df_rv_lines.to_csv("data/rv_lines.csv", index=None)
df_rv_lines.head()

## Validation

### Checking for missing roots

import numpy as np

def print_roots_attestation_info(df_roots, df_rv_lines, present_class):
    # our starting list of roots
    roots_initial = np.sort(df_roots.query(
        f"present_class == '{present_class}'"
    )["root"].unique())
    print(f"Starting list ({len(roots_initial)}):\n{roots_initial}\n")
    
    # attested roots
    roots_present = np.sort(df_rv_lines.query(
        f"present_class == '{present_class}'"
    )["root"].unique())
    print(f"Attested ({len(roots_present)}):\n{roots_present}\n")

    # missing roots
    roots_absent = np.sort(
        np.setdiff1d(roots_initial, roots_present)
    )
    print(f"Missing ({len(roots_absent)}):\n{roots_absent}\n")

print_roots_attestation_info(df_roots, df_rv_lines, "ninth")

print_roots_attestation_info(df_roots, df_rv_lines, "fifth")

TODO explain that these roots not being in RV is ok (and expected):
    
* Some roots are attested only as weak stem before vowel (i.e no _nī_ form). e.g.:  [_uṣ_](https://vedaweb.uni-koeln.de/rigveda/view/id/09.097.39)
* Some roots are there in later vedic texts.
* Some roots are already marked as later language in whitney (to be found only in Classical / Epic sanskrit

TODO also try searching by alternate stem/root vowel forms for the missing roots (and also for the attested?), to see if there are actually recorded in those forms

### Checking found word against padapatha text

# from src/lib/meter.py
from src.lib.meter import clean_lubotsky_padapatha

mismatches = []

no_of_mismatches = 0

for line in rv_lines_annotated:
    padapatha_text_cleaned = clean_lubotsky_padapatha(line["text_padapatha"])
    padapatha_parts = padapatha_text_cleaned.split(' ')
    
    word_from_padapatha = padapatha_parts[line["word_position_no"] - 1]
    
    if line["word"] != word_from_padapatha:
        no_of_mismatches += 1
        print(
            f"{line['line_no']}:{line['word_position_no']}", 
            f"({line['stem']}) {line['word']} ≠ {word_from_padapatha}",
            #f"[{padapatha_text_cleaned}]",
            f"[{line['text_padapatha']}]"
        )
        
print(f"\nFound {no_of_mismatches} mismatches.")

Ignore cases of final vowel lengthening above (especially with imperatives) -- they are not really mismatches.

TODO look at vowel positions for these? but not really in the scope of our investigation

## Summary of Results

import pandas
res_df_rv_lines = pandas.read_csv("data/rv_lines.csv", keep_default_na=False)
res_df_rv_lines.head()

res_df_rv_lines \
    .groupby(["present_class", "stem_type"]) \
    .size().to_frame("count").reset_index() \
    .sort_values(["present_class", "stem_type"], ascending=[False, False])
